Create print.rs
