+++ src/main.rs
@@ -1,6 +1,10 @@
 extern crate rand;
+mod generate;
+mod solve;
+mod print;
 
 use rand::{thread_rng, Rng};
+use print::print_to_console;
 
 static SIZE: usize = 4;
 
@@ -25,6 +29,10 @@ fn reset_maze(maze: Vec<Vec<Vec<i32>>>) -> Vec<Vec<Vec<i32>>> {
         }
         solution.push(new_row);
     }
+
+    solution[0][0][4] = 2;
+    solution[SIZE - 1][SIZE - 1][3] = 1;
+
     solution
 }
 
@@ -73,11 +81,6 @@ fn generate_maze() -> Vec<Vec<Vec<i32>>> {
             c = popped.1;
         }
     }
-
-    // set start
-    grid[0][0][4] = 2;
-    // set finish
-    grid[SIZE - 1][SIZE - 1][3] = 1;
 
     grid
 }
@@ -186,52 +189,6 @@ fn manhattan_dist(a: (i32, i32), b: (i32, i32)) -> i32 {
     let y_val = (a.0 - b.0).abs();
     let x_val = (a.1 - b.1).abs();
     x_val + y_val
-}
-
-fn print_to_console(grid: &Vec<Vec<Vec<i32>>>) {
-    println!("");
-    for row in 0..grid.len() {
-        if row == 0 { print_top_row(&grid[row]); };
-        print_vertical_walls(&grid[row]);
-        print_horizontal_walls(&grid[row]);
-    }
-    println!("");
-}
-
-fn print_top_row(row: &Vec<Vec<i32>>) {
-    let mut top = vec!["+"];
-    for cell in row {
-        if cell[4] == 2 {
-            top.push("   +");
-        } else {
-            top.push("---+");
-        }
-    }
-    let joined = top.join("");
-    println!("{:?}", joined);
-}
-
-fn print_vertical_walls(row: &Vec<Vec<i32>>) {
-    let mut vertical = vec!["|"];
-    for cell in row {
-        if cell[4] == 1 { vertical.push(" * "); } else { vertical.push("   "); }
-        if cell[2] == 1 { vertical.push(" "); } else { vertical.push("|"); }
-    }
-    let joined = vertical.join("");
-    println!("{:?}", joined);
-}
-
-fn print_horizontal_walls(row: &Vec<Vec<i32>>) {
-    let mut horizontal = vec!["+"];
-    for cell in row {
-        if cell[3] == 1 {
-            horizontal.push("   +");
-        } else {
-            horizontal.push("---+");
-        }
-    }
-    let joined = horizontal.join("");
-    println!("{:?}", joined);
 }
 
 #[cfg(test)]
